moved tests to test_relationship.py
